UPDATE
UCharisma divided into UCharismaAPI and UPlaythrough, as required at the "Unreal  SDK Improvement" document.
